revert to minimal examples for the OpenAPI and use subset for demo env
1. Package the minimal examples in fhir-openapi. The Liberty OpenAPI UI
doesn't seem to like our complete-mock examples. For example, instead of
including the datetime values it just uses an empty object for these
fields (which fails FHIR validation). It seems to do a little better
with our minimal examples (although these are a tad less useful in terms
of documenting the allowed fields). Perhaps more importantly, switching
back to the minimal examples saves about 50,000 lines in the all-in-one
openapi definition, which should make it a bit more management.

2. Overwrite the default openapi.json in our demo environment with one
that matches the subset of resources that are actually exposed in that
environment. This makes viewing and playing with the API surface area a
bit more managable and prevents users from tryinig to use endpoints
which aren't actually suppoted in this environment.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>